+++ onTrack.API/client/api/kpi/kpiController.js
@@ -8,18 +8,30 @@
 			$scope.deleteKpi = function() {
 				var token = $window.sessionStorage['jwt']
 
-				$http.delete('/api/clients/' + $stateParams['clientid']
-				 + '/kpis/' + $stateParams['kpiid'], {
-					headers: {
-						'Authorization': `Bearer ${token}`
-					}
-				})
-				.success(function(data){
-					var clientId = {'id': $stateParams['clientid'] + ''}
-					$state.go('client',clientId )
-				})
-				.error(function(err) {
-					console.log(err)
+
+				swal({
+				  title: "Are you sure?",
+				  text: "You will not be able to recover this client!",
+				  type: "warning",
+				  showCancelButton: true,
+				  confirmButtonColor: "#DD6B55",
+				  confirmButtonText: "Yes, delete it!",
+				  closeOnConfirm: true,
+				  html: false
+				}, function(){
+					$http.delete('/api/clients/' + $stateParams['clientid']
+					 + '/kpis/' + $stateParams['kpiid'], {
+						headers: {
+							'Authorization': `Bearer ${token}`
+						}
+					})
+					.success(function(data){
+						var clientId = {'id': $stateParams['clientid'] + ''}
+						$state.go('client',clientId )
+					})
+					.error(function(err) {
+						console.log(err)
+					})
 				})
 			}
